requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
cachelib==0.3.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
distlib==0.3.3
filelock==3.1.0
Flask==2.0.1
Flask-Session==0.4.0
greenlet==1.1.1
gunicorn==20.1.0
helpers==0.2.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
multidict==5.1.0
mysql-connector==2.2.9
pipenv==2021.5.29
platformdirs==2.4.0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tempfile2==0.1.1
termcolor==1.1.0
typing-extensions==3.10.0.2
urllib3==1.26.7
virtualenv==20.8.1
virtualenv-clone==0.5.7
Werkzeug==2.0.1
yarl==1.6.3
